initial work to generify functions and userdata types
the rationale is that typically functions receive a well-defined list of argument types
and require either an ArgumentIterator to peel out the values or directly cast their
parameters to the specific type.

Not only is that error prone, it also required duplicate work all over the place.
By allowing to specify the argument types no such casting is required for the common case
as the simple type parameter allows IDEs to properly create method signatures.

The same applies to userdata, which is very often used to expose Java classes into Lua.
In those cases the types are well known to the programmer and no helper fields, type casts or
other checks are required, because the user data is already encapsulated.

Thus the boilerplate required is reduced and the user code becomes clearer.